requirements.txt
PyYAML
jsonschema
python-dotenv
requests
rich
build